test: Move unit tests to tests folder
