+++ contracts/oracle/wasm/src/price_source.rs
@@ -210,51 +210,3 @@ impl PriceSourceChecked<Empty> for WasmPriceSourceChecked {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use mars_testing::mock_dependencies;
-
-    use super::*;
-
-    #[test]
-    fn display_fixed_price_source() {
-        let ps = WasmPriceSource::Fixed {
-            price: Decimal::from_ratio(1u128, 2u128),
-        };
-        assert_eq!(ps.to_string(), "fixed:0.5")
-    }
-
-    #[test]
-    fn display_spot_price_source() {
-        let ps = WasmPriceSourceChecked::AstroportSpot {
-            pair_address: Addr::unchecked("fake_addr"),
-            route_assets: vec![],
-        };
-        assert_eq!(ps.to_string(), "astroport_spot:fake_addr. Route: ")
-    }
-
-    #[test]
-    fn display_spot_price_source_with_route() {
-        let ps = WasmPriceSourceChecked::AstroportSpot {
-            pair_address: Addr::unchecked("fake_addr"),
-            route_assets: vec!["fake_asset1".to_string(), "fake_asset2".to_string()],
-        };
-        assert_eq!(ps.to_string(), "astroport_spot:fake_addr. Route: fake_asset1,fake_asset2")
-    }
-
-    // TODO: Display test for twap
-
-    #[test]
-    fn validate_fixed_price_source() {
-        let ps = WasmPriceSource::Fixed {
-            price: Decimal::from_ratio(1u128, 2u128),
-        };
-        let deps = mock_dependencies(&[]);
-        let price_sources = Map::new("price_sources");
-        let denom = "uusd";
-        let base_denom = "uusd";
-        let res = ps.validate(&deps.as_ref(), denom, base_denom, &price_sources);
-        assert!(res.is_ok());
-    }
-}
